+++ GravatarApp/CommonViews/CheckmarkCircleView.swift
