+++ GravatarApp/CommonViews/DimmingActivityIndicator.swift
