+++ GravatarApp/CommonViews/HeaderAvatarView.swift
